+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #17 Add pollable paginated ServiceRequest endpoint for active filler orders
+- #15 Create FHIR instrument ServiceRequests when an instrument is assigned to an analysis
 - #14 Expose SENAITE instruments as FHIR Device resources
 - #11 Build Specimen resources from SENAITE Analysis Request data
 - #13 Fix dropped marital status on FHIR patient import